third_party/blink/web_tests/fast/dom/shadow/access-key-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests to ensure that accesskey works in regard to shadow DOM boundary. Can only run within DRT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dispatchedEvent("focus") is ["inputE"]
PASS dispatchedEvent("focus") is ["inputE"]
PASS dispatchedEvent("focus") is ["inputE"]
PASS successfullyParsed is true

TEST COMPLETE
